Introduced ChangeEvent to decouple changes from specific documents
Currently we rely on signals to notify the UI about specific changes
that can be made to objects that are part of a document. The MapDocument
and TilesetDocument both expose a different set of signals.

One problem with this is that MapObjects can be part of both types of
document, since they are also part of the tile collision shapes that are
stored in a tileset document.

By having only a single "changed" signal and a separate hierarchy of
ChangeEvents, we can decouple the change events from the documents and
allow for example map object changes to be sent by both document types.
Hopefully this change will allow making the Tile Collision Editor less
hacky and enable scripts access to the tile collision shapes.