GmxPlugin: Fixed compatibility with GameMaker 1.4.9999
A new "colour" attribute was added at some point, which causes compile
errors in GameMaker when it is missing when trying to Run the game.
It is now written out as pure white, the default. The alpha component of
the color is set to the effective opacity of the layer.

I've also added the "locked" attribute, setting it to -1 if a tile or
instance is part of a locked layer.