Added option to change object selection behavior
Previously when clicking objects it would always take the top-most object
from any visible and unlocked layer.

This new option allows changing that such that when available, an object
from one of the currently selected layers is preferred. It can also be
set to only do this, when the highlighting of current layer option is
enabled (the new default).